// lae_desktop/src/app/models/parent-model.ts
export class ParentModel{
    parentId: number;
    parentFirstName: string;
    parentLastname: string;
    parentTitle: string;
    parentSex: string;
    parentTels: string;
    parentEmail: string;
    parentReligion: string;
    parentAddress: string;
    createdBy: string;
    dateCreated: string;
    modifiedBy: string;
    dateModified: string;
    
}